requirements.txt
flask>=2.0.1
numpy>=1.22.0
tensorflow>=2.12.0
opencv-python-headless>=4.5.3.56
werkzeug>=2.0.1
google-auth>=2.3.3
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.31.0
gunicorn>=20.1.0
gdown>=4.6.0
requests>=2.28.1
Pillow>=8.3.2
scikit-learn>=1.0.2
matplotlib>=3.5.0
